apmsetup.js -> elastic-apm-node/start-next.js; drop the .env file handling, will update docs to suggest elastic-apm-node.js
