+++ examples/nextjs/.gitignore
@@ -1,5 +1,7 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
+/elastic-apm-node.js
+
 # dependencies
 /node_modules
 /.pnp
