+++ package.json
@@ -15,7 +15,7 @@
     "lint:yaml-files": "./dev-utils/lint-yaml-files.sh  # requires node >=10",
     "coverage": "COVERAGE=true ./test/script/run_tests.sh",
     "test": "./test/script/run_tests.sh",
-    "test:deps": "dependency-check start.js index.js 'lib/**/*.js' 'test/**/*.js' '!test/instrumentation/modules/next/a-nextjs-app' --no-dev -i async_hooks -i perf_hooks -i parseurl -i node:http",
+    "test:deps": "dependency-check index.js start.js start-next.js 'lib/**/*.js' 'test/**/*.js' '!test/instrumentation/modules/next/a-nextjs-app' --no-dev -i async_hooks -i perf_hooks -i parseurl -i node:http",
     "test:tav": "tav --quiet && (cd test/instrumentation/modules/next/a-nextjs-app && tav --quiet)",
     "test:docs": "./test/script/docker/run_docs.sh",
     "test:types": "tsc --project test/types/tsconfig.json && tsc --project test/types/transpile/tsconfig.json && node test/types/transpile/index.js && tsc --project test/types/transpile-default/tsconfig.json && node test/types/transpile-default/index.js",
@@ -38,8 +38,10 @@
     "lib",
     "types",
     "start.js",
+    "start-next.js",
     "index.d.ts",
-    "start.d.ts"
+    "start.d.ts",
+    "start-next.d.ts"
   ],
   "repository": {
     "type": "git",
