+++ start-next.d.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright Elasticsearch B.V. and other contributors where applicable.
+ * Licensed under the BSD 2-Clause License; you may not use this file except in
+ * compliance with the BSD 2-Clause License.
+ */
+
+import * as agent from './';
+export = agent;
